add ram.py
